Refactoring + finishing adding tests
